[feat] Character의 카메라를 PlayerCharacter로 이전
