+++ EngineSIU/EngineSIU/Engine/Source/Games/LastWar/Characters/EnemyCharacter.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "GameFramework/Character.h"
 
-class AEnemy : public ACharacter
+class AEnemyCharacter : public ACharacter
 {
 };
