--- EngineSIU/EngineSIU/Engine/Source/Games/LastWar/Characters/Player.cpp
@@ -1,1 +1,0 @@
-#include "Player.h"
