--- EngineSIU/EngineSIU/Engine/Source/Games/LastWar/Characters/Player.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "GameFramework/Character.h"
-
-class APlayer : public ACharacter
-{
-};
-
