+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/GameFramework/Character.h
@@ -4,7 +4,7 @@
 class UInputComponent;
 class AController;
 class UStaticMeshComponent;
-class UCameraComponent;
+
 class ACharacter : public APawn
 {
     DECLARE_CLASS(ACharacter, APawn)
@@ -18,26 +18,11 @@ public:
     virtual void PossessedBy(AController* NewController) override;
     virtual void UnPossessed() override;
 
-    // === 이동 관련 ===
-    void MoveForward(float Value);
-    void MoveRight(float Value);
-    void MoveLeft(float Value);
-    void MoveBackward(float Value);
-
-    // === 카메라 관련 ===
-    UCameraComponent* GetFollowCamera() const { return FollowCamera; }
-
-private:
+protected:
     UPROPERTY
     (USceneComponent*, RootScene, = nullptr);
 
     UPROPERTY
     (UStaticMeshComponent*, BodyMesh, = nullptr);
-
-    //UPROPERTY
-    //(USpringArmComponent*, CameraBoom, = nullptr);
-
-    UPROPERTY
-    (UCameraComponent*, FollowCamera, = nullptr);
 };
