+++ EngineSIU/EngineSIU/EngineSIU.vcxproj
@@ -171,7 +171,8 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="Engine\Source\Editor\UnrealEd\EditorConfigManager.cpp" />
-    <ClCompile Include="Engine\Source\Games\LastWar\Characters\Player.cpp" />
+    <ClCompile Include="Engine\Source\Games\LastWar\Characters\EnemyCharacter.cpp" />
+    <ClCompile Include="Engine\Source\Games\LastWar\Characters\PlayerCharacter.cpp" />
     <ClCompile Include="Engine\Source\Runtime\Core\Math\CollisionMath.cpp" />
     <ClCompile Include="Engine\Source\Runtime\Engine\Classes\Components\LuaScriptComponent.cpp" />
     <ClCompile Include="Engine\Source\Runtime\Engine\Classes\Components\Shapes\BoxComponent.cpp" />
@@ -402,7 +403,8 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Engine\Source\Games\LastWar\Characters\Player.h" />
+    <ClInclude Include="Engine\Source\Games\LastWar\Characters\EnemyCharacter.h" />
+    <ClInclude Include="Engine\Source\Games\LastWar\Characters\PlayerCharacter.h" />
     <ClInclude Include="Engine\Source\Runtime\Core\Math\CollisionMath.h" />
     <ClInclude Include="Engine\Source\Editor\UnrealEd\EditorConfigManager.h" />
     <ClInclude Include="Engine\Source\Runtime\Engine\Classes\Components\LuaScriptComponent.h" />
