+++ EngineSIU/EngineSIU/editor.ini
@@ -1,6 +1,6 @@
-ScenePath=C:\Users\Jungle\Desktop\TechLab\Week8\GTL_W08_T5\EngineSIU\EngineSIU\Saved\Sponza.scene
+ScenePath=Saved/DefaultLevel.scene
 SplitterH.SplitRatio=0.500000
-PerspectiveCameraLocX3=-8.234695
+PerspectiveCameraLocX3=-17.492722
 PerspectiveCameraRotY1=0.000000
 SplitterV.SplitRatio=0.500000
 PerspectiveCameraLocY0=-2.476464
@@ -26,7 +26,7 @@ ViewportType0=0
 CameraSpeedScalar1=1.000000
 GridSize1=10.000000
 PerspectiveCameraLocX1=0.000000
-PerspectiveCameraRotY3=27.900068
+PerspectiveCameraRotY3=39.799965
 PerspectiveCameraLocY1=0.000000
 ScreenHeight=961
 PerspectiveCameraLocZ1=0.000000
@@ -46,9 +46,9 @@ ScreenWidth=1384
 ViewportType2=3
 CameraSpeedScalar3=0.402002
 GridSize3=10.000000
-PerspectiveCameraLocY3=-3.832469
-PerspectiveCameraLocZ3=4.566734
-PerspectiveCameraRotZ3=-346.202942
+PerspectiveCameraLocY3=1.699071
+PerspectiveCameraLocZ3=21.589838
+PerspectiveCameraRotZ3=-360.103821
 ViewMode3=2
 ViewportType3=0
 bMultiView=0
